+++ src/main.rs
@@ -311,15 +311,17 @@ fn main() {
             namespace
         }
     };
-    edebugln!("Found context={context} namespace={namespace}");
+    edebugln!("Found context={context:?} namespace={namespace:?}");
 
     // Rebuild command
     let command = env::args().skip(1).collect::<Vec<String>>().join(" ");
-    edebugln!("Received command={command}");
+    edebugln!("Received command={command:?}");
 
     // Figure out what to do
     let (validation, record, amendment, reason) = identify_actions(&context, &command, include, exclude);
-    edebugln!("Decided validation={validation} record={record} amendment={amendment} reason={reason}");
+    edebugln!(
+        "Decided validation={validation:?} record={record:?} amendment={amendment:?} reason={reason:?}"
+    );
 
     // Set new context if needed
     if validation {