+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Use official markdownlint github action
+- Use variable debug format in debug logs
 
 ## v2.3.0