+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## v2.5.0
+
 - Apply clippy lint suggestions on macro rules
 - Bump markdownlint-cli2 action from 7 to 8 (<https://github.com/badouralix/kubekeeper/pull/3>)
 - Update magic to detect native kubectl commands (<https://github.com/badouralix/kubekeeper/pull/5>)